Refactor DemoSearchInput to streamline search entry creation and enhance code organization
